new version to spider and keywords
